+++ CotEditor/Sources/DocumentViewController.swift
@@ -69,7 +69,7 @@ final class DocumentViewController: NSSplitViewController, NSMenuItemValidation,
         
         super.viewDidLoad()
         
-        // setup status bar
+        // set user defaults
         let defaults = UserDefaults.standard
         self.isStatusBarShown = defaults[.showStatusBar]
         self.showsInvisibles = defaults[.showInvisibles]
@@ -77,6 +77,16 @@ final class DocumentViewController: NSSplitViewController, NSMenuItemValidation,
         self.showsNavigationBar = defaults[.showNavigationBar]
         self.wrapsLines = defaults[.wrapLines]
         self.showsPageGuide = defaults[.showPageGuide]
+        
+        // set writing direction
+        switch WritingDirection(defaults[.writingDirection]) {
+        case .leftToRight:
+            break
+        case .rightToLeft:
+            self.writingDirection = .rightToLeft
+        case .vertical:
+            self.verticalLayoutOrientation = true
+        }
         
         // set theme
         let themeName = ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
@@ -524,7 +534,11 @@ final class DocumentViewController: NSSplitViewController, NSMenuItemValidation,
     @objc var verticalLayoutOrientation: Bool {
         
         get {
-            return self.document?.isVerticalText ?? false
+            guard let textView = self.focusedTextView else {
+                return WritingDirection(UserDefaults.standard[.writingDirection]) == .vertical
+            }
+            
+            return textView.layoutOrientation == .vertical
         }
         
         set {